+++ src/themes/elegant/404.php
@@ -23,7 +23,7 @@ outputNavbar($wiki);
 outputBanner($wiki);
 
 ?>
-<section class="section-main container">
+<section class="section-main container page-404">
   <div class="row">
     <div class="col-12">
       <?php if ($wiki->core->mayCreatePath()) { ?>