+++ src/themes/elegant/error.php
@@ -23,7 +23,7 @@ outputNavbar($wiki);
 outputBanner($wiki);
 
 ?>
-<section class="section-main container">
+<section class="section-main container page-error">
   <div class="row">
     <div class="col-12">
       <p><?php __('Sorry, an error occured (or you don\'t have permission to do this).'); ?></p>